[refactor] Modulization (ProfileSection)
ProfileSection에서 ProfileImageUploader 분리
